database/pyproject.toml
[project]
name = "database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.117.1",
    "greenlet>=3.2.4",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
